don't use context for auth cookie
