+++ src/hooks/useAuth.tsx
@@ -1,11 +1,9 @@
 import { isNull } from '@fullstacksjs/toolbox';
-import { useContext } from 'react';
 
-import { AuthCtx } from '../contexts/AuthCtx';
+import { getClientCookie } from '../infra';
 
 export const useAuth = () => {
-  const ctx = useContext(AuthCtx);
-  // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
-  if (isNull(ctx)) throw new Error('Auth Context Provider Not Found');
-  return ctx;
+  const cookie = getClientCookie();
+  const isAuth = isNull(cookie) ? false : cookie.isAuth;
+  return { isAuth };
 };
